Fixed bugs, changed animal data structure and changed animalform.
Created new object Animal to fill with each animal data. This object will be used in the form to be filled. Added functionallity to write to nfc once completed editing form (needs to be checked, no snackbar if error). Changed database to json type. Improved animalform so it can be edited without errors, data now can be selected from datapicker.